src/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3  # 必须使用2.0.x版本
Flask-SocketIO==5.3.2
python-engineio==4.3.4
python-socketio==5.7.2
eventlet==0.33.3
gunicorn==20.1.0